write a status file

+++ startup.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
+
+if [ -f /local/startup_service_done ]; then
+    exit 0
+fi
+
 # install NVIDIA drivers through ubuntu-drivers
 apt update
 apt install -y ubuntu-drivers-common
@@ -8,3 +13,4 @@ ubuntu-drivers install --gpgpu
 version=$(dpkg-query --show nvidia-kernel-common-* | awk '{print $1}' | awk -F'-' '{print $4"-" $5}')
 apt install -y nvidia-utils-$version nvidia-driver-$version
 
+> /local/startup_service_done
